+++ api/core/v3/entity_state.pb.go
@@ -524,10 +524,7 @@ func (m *EntityState) Unmarshal(dAtA []byte) error {
 			if err != nil {
 				return err
 			}
-			if skippy < 0 {
-				return ErrInvalidLengthEntityState
-			}
-			if (iNdEx + skippy) < 0 {
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
 				return ErrInvalidLengthEntityState
 			}
 			if (iNdEx + skippy) > l {