+++ CHANGELOG.md
@@ -24,6 +24,7 @@ for controlling agent retry exponential backoff behaviour. --retry-min and
 a decimal multiplier value.
 - Added ProcessedBy field to check results. The ProcessedBy field indicates which
 agent processed a particular event.
+- Added `core/v3.Pipeline` resource for configuring Pipeline resources.
 
 ### Changed
 - When keepalived encounters round-robin ring errors, the backend no longer